# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JARVIS AI Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthRequest {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface ChatRequest {
  message: string;
  /** @nullable */
  sessionId?: string | null;
}

export interface ChatResponse {
  reply: string;
  sessionId: string;
  timestamp: string;
}

export interface VoiceCommandRequest {
  command: string;
}

/**
 * @nullable
 */
export type VoiceCommandResponseData = { [key: string]: unknown } | null;

export interface VoiceCommandResponse {
  action: string;
  reply: string;
  /** @nullable */
  data?: VoiceCommandResponseData;
  timestamp: string;
}

export interface PlayMusicRequest {
  songName: string;
}

export interface PlayMusicResponse {
  songName: string;
  url: string;
  found: boolean;
}

export interface SongEntry {
  name: string;
  url: string;
}

export interface MusicLibraryResponse {
  songs: SongEntry[];
  total: number;
}

export interface NewsArticle {
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  source: string;
  publishedAt: string;
}

export interface NewsResponse {
  articles: NewsArticle[];
  total: number;
}

export interface CommandHistoryEntry {
  id: number;
  type: string;
  input: string;
  response: string;
  /** @nullable */
  action?: string | null;
  createdAt: string;
}

export interface HistoryResponse {
  entries: CommandHistoryEntry[];
  total: number;
}

export interface CommandCount {
  action: string;
  count: number;
}

export interface ActivityStats {
  totalCommands: number;
  totalChats: number;
  totalVoiceCommands: number;
  totalMusicPlayed: number;
  totalNewsRequests: number;
  mostUsedCommands: CommandCount[];
}

export type GetNewsParams = {
  country?: string;
  limit?: number;
};

export type GetHistoryParams = {
  limit?: number;
};