# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JARVIS AI Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const RegisterBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

/**
 * @summary Login with credentials
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Get current user info
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Send a text message to JARVIS
 */
export const SendChatBody = zod.object({
  message: zod.string(),
  sessionId: zod.string().nullish(),
});

export const SendChatResponse = zod.object({
  reply: zod.string(),
  sessionId: zod.string(),
  timestamp: zod.coerce.date(),
});

/**
 * @summary Process a voice command (text transcript)
 */
export const VoiceCommandBody = zod.object({
  command: zod.string(),
});

export const VoiceCommandResponse = zod.object({
  action: zod.string(),
  reply: zod.string(),
  data: zod.object({}).passthrough().nullish(),
  timestamp: zod.coerce.date(),
});

/**
 * @summary Get a music link by song name
 */
export const PlayMusicBody = zod.object({
  songName: zod.string(),
});

export const PlayMusicResponse = zod.object({
  songName: zod.string(),
  url: zod.string(),
  found: zod.boolean(),
});

/**
 * @summary Get all available songs in the library
 */
export const GetMusicLibraryResponse = zod.object({
  songs: zod.array(
    zod.object({
      name: zod.string(),
      url: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get top news headlines
 */
export const getNewsQueryCountryDefault = `us`;
export const getNewsQueryLimitDefault = 10;

export const GetNewsQueryParams = zod.object({
  country: zod.coerce.string().default(getNewsQueryCountryDefault),
  limit: zod.coerce.number().default(getNewsQueryLimitDefault),
});

export const GetNewsResponse = zod.object({
  articles: zod.array(
    zod.object({
      title: zod.string(),
      description: zod.string().nullish(),
      url: zod.string(),
      source: zod.string(),
      publishedAt: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get command history
 */
export const getHistoryQueryLimitDefault = 50;

export const GetHistoryQueryParams = zod.object({
  limit: zod.coerce.number().default(getHistoryQueryLimitDefault),
});

export const GetHistoryResponse = zod.object({
  entries: zod.array(
    zod.object({
      id: zod.number(),
      type: zod.string(),
      input: zod.string(),
      response: zod.string(),
      action: zod.string().nullish(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get activity statistics
 */
export const GetActivityStatsResponse = zod.object({
  totalCommands: zod.number(),
  totalChats: zod.number(),
  totalVoiceCommands: zod.number(),
  totalMusicPlayed: zod.number(),
  totalNewsRequests: zod.number(),
  mostUsedCommands: zod.array(
    zod.object({
      action: zod.string(),
      count: zod.number(),
    }),
  ),
});