# lib/api-zod/src/generated/types/historyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JARVIS AI Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { CommandHistoryEntry } from "./commandHistoryEntry";

export interface HistoryResponse {
  entries: CommandHistoryEntry[];
  total: number;
}