# lib/api-zod/src/generated/types/musicLibraryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JARVIS AI Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { SongEntry } from "./songEntry";

export interface MusicLibraryResponse {
  songs: SongEntry[];
  total: number;
}