# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JARVIS AI Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsArticle {
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  source: string;
  publishedAt: string;
}