# lib/api-zod/src/generated/types/newsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JARVIS AI Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { NewsArticle } from "./newsArticle";

export interface NewsResponse {
  articles: NewsArticle[];
  total: number;
}