# lib/api-zod/src/generated/types/voiceCommandResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JARVIS AI Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { VoiceCommandResponseData } from "./voiceCommandResponseData";

export interface VoiceCommandResponse {
  action: string;
  reply: string;
  /** @nullable */
  data?: VoiceCommandResponseData;
  timestamp: Date;
}